+++ dev/user.clj
@@ -1,13 +1,50 @@
 (ns user
   (:require
    [clj-reload.core :as reload]
-   [java-time.api :as jt]
-   [taoensso.telemere :as tel]
-   [hkimjp.datascript :as ds]
-   [hkimjp.jpy.system :refer [start-system stop-system]]))
+   [environ.core :refer [env]]
+   [org.httpkit.server :as hk]
+   [ring.middleware.reload :refer [wrap-reload]]
+   [taoensso.telemere :as t]
+   [hkimjp.jpy.routes :refer [root-handler]]
+   [hkimjp.datascript :as ds]))
 
+;--------------------------
+; ring-devel
+(defonce server (atom nil))
+
+(defn start-server []
+  (let [handler (wrap-reload #'root-handler)]
+    (let [port (parse-long (or (env :port) "3000"))]
+      (reset! server (hk/run-server handler {:port port}))
+      (println (str "http-kit started at port " port)))))
+
+(defn stop-server []
+  (@server)
+  (reset! server nil))
+
+(defn start-system []
+  (t/log! {:level :info
+           :id "start-system"
+           :msg (env :develop)
+           :data {:datascript (env :datascript)}})
+  (try
+    (ds/start-or-restore {:url (env :datascript)})
+    (start-server)
+    (catch Exception e
+      (t/log! :fatal (.getMessage e))
+      (System/exit 0))))
+
+(defn stop-system []
+  (stop-server)
+  (ds/stop))
+
+(start-system)
+; (stop-system)
+
+;---------------------------
+; clj-reload
 (reload/init
- {:dirs ["src" "dev" "test"]
+ {:dirs ["src" "test"]
   :no-reload '#{user}
   :unload-hook 'before-unload
   :after-reload 'start-system})
@@ -19,39 +56,4 @@
   (start-system))
 
 ; (reload/reload)
-
-(tel/set-min-level! :info)
-
-(start-system)
-
-; -------------------------
-(comment
-  ;; initialize
-  (ds/put! {:num 0
-            :problem "dummy problem"
-            :stat "ok"
-            :datetime (jt/local-date-time)})
-
-  (ds/put! {:current 0})
-  :rcf)
-
-(comment
-  (ds/qq '[:find ?e ?num
-           :where
-           [?e :current ?num]])
-
-  (ds/pl 1)
-
-  (ds/qq '[:find ?problem
-           :where
-           [?e :num 19]
-           [?e :problem ?problem]])
-
-  (ds/qq '[:find ?e
-           :where
-           [?e _ _]])
-
-  (ds/pl 5)
-
-  :rcf)