+++ src/hkimjp/jpy/system.clj
@@ -1,26 +1,20 @@
 (ns hkimjp.jpy.system
   (:require
    [environ.core :refer [env]]
-   [ring.adapter.jetty :as jetty]
+   [org.httpkit.server :as hk]
    [taoensso.telemere :as t]
-   [hkimjp.jpy.routes :as routes]
+   [hkimjp.jpy.routes :refer [root-handler]]
    [hkimjp.datascript :as ds]))
 
 (defonce server (atom nil))
 
-(defn start-jetty
-  []
-  (let [port (parse-long (or (env :port) "3000"))
-        handler (if (some? (env :develop))
-                  #'routes/root-handler
-                  routes/root-handler)]
-    (reset! server (jetty/run-jetty handler {:port port :join? false}))
-    (t/log! :info (str "server started at port " port))))
+(defn start-server []
+  (let [port (parse-long (or (env :port) "3000"))]
+    (reset! server (hk/run-server root-handler {:port port}))))
 
 (defn stop-server []
-  (when @server
-    (.stop @server)
-    (t/log! :info "server stopped.")))
+  (@server)
+  (reset! server nil))
 
 (defn start-system []
   (t/log! {:level :info
@@ -29,7 +23,7 @@
            :data {:datascript (env :datascript)}})
   (try
     (ds/start-or-restore {:url (env :datascript)})
-    (start-jetty)
+    (start-server)
     (catch Exception e
       (t/log! :fatal (.getMessage e))
       (System/exit 0))))