+++ src/hkimjp/jpy/workspace.clj
@@ -5,71 +5,75 @@
    [ring.util.anti-forgery :refer [anti-forgery-field]]
    [taoensso.telemere :as tel]
    [hkimjp.datascript :as ds]
+   [hkimjp.jpy.answers :as answers]
    [hkimjp.jpy.util :refer [btn user]]
    [hkimjp.jpy.view :refer [page error-page hx redirect]]))
 
-(defn answer [{{:keys [e]} :path-params}]
+(defn answer
+  "called from answers-section, returns hx response"
+  [{{:keys [e]} :path-params}]
   (tel/log! {:level :info :id "answer" :msg e})
-  (hx [:pre (:answer (ds/pl (parse-long e)))]))
+  (let [{:keys [answer p/id]} (ds/pl (parse-long e))
+        {:keys [problem]} (ds/pl id)]
+    (hx [:div
+         [:p problem]
+         [:pre answer]])))
 
-(def list-answers-q
-  '[:find ?e ?num
-    :in $ ?author
-    :where
-    [?e :login ?author]
-    [?e :p/num ?num]])
+(comment
+  (let [{:keys [answer p/id]} (ds/pl 4)]
+    id)
+  (ds/pl 4)
+  (ds/pl 3)
+  :rcf)
 
-(defn list-answers-by [author]
+(defn answers-section [author]
   [:div
    [:div.font-bold "answers"]
    (into
     [:div.flex.gap-x-4]
-    (for [[e num] (->> (ds/qq list-answers-q author) (sort-by first))]
+    (for [[e num] (->> (answers/list-answers author) (sort-by first))]
       [:a.underline {:hx-get (format "/workspace/answer/%d" e)
                      :hx-target "#answer"} num]))
    [:div#answer.my-4]])
 
-;(list-answers-by "hkimura")
-
-(def ^:private current-problem-id
-  '[:find ?e ?current
-    :where
-    [?e :current ?current]])
-
 (def ^:private current-problem
-  '[:find [?e ?num ?problem]
-    :in $ ?num
+  '[:find [?id ?problem]
     :where
-    [?e :num ?num]
-    [?e :problem ?problem]])
+    [?e :current ?id]
+    [?id :problem ?problem]])
 
+; (ds/qq current-problem)
 (defn index [request]
   (let [author (user request)
-        [_ id] (apply max-key first (ds/qq current-problem-id))
-        [e num problem] (ds/qq current-problem id)]
-    (tel/log! {:level :info :id "index" :data {:e e :num num :p problem}})
+        [id problem] (ds/qq current-problem)]
+    (tel/log! {:level :info :id "index" :data {:id id :problem problem}})
     (page
      [:div.m-4
       [:div.text-2xl.font-medium "workspace"]
-      [:div.flex.gap-x-4 [:div num] [:div problem]]
+      [:div.flex.gap-x-4 [:div id] [:div problem]]
       [:form {:method "post"}
        (h/raw (anti-forgery-field))
        [:input {:type "hidden" :name "login" :value author}]
-       [:input {:type "hidden" :name "num" :value num}]
+       [:input {:type "hidden" :name "id" :value id}]
        [:textarea {:class "w-full h-64 border-1 p-2"
                    :name "answer"
                    :placeholder "your answer, please."}]
        [:button {:class btn} "submit"]]
       [:br]
-      (list-answers-by author)])))
+      (answers-section author)])))
 
-(defn upload! [{{:keys [login num answer]} :params :as request}]
+; (index {})
+
+;; shoule be in answer.clj?
+(defn upload! [{{:keys [login id answer]} :params :as request}]
   (tel/log! {:level :info
              :id "upload!"
              :data (dissoc (:params request) :__anti-forgery-token)})
   (try
-    (ds/put! {:login login :p/num num :answer answer :datetime (jt/local-date-time)})
+    (ds/put! {:login login
+              :p/id (parse-long id)
+              :answer answer
+              :datetime (jt/local-date-time)})
     (redirect "/workspace")
     (catch Exception e
       (error-page (.getMessage e)))))
-